Move to the separate test case
