+++ ci/it/integration_test.go
@@ -59,3 +59,8 @@ func TestABTestcase(t *testing.T) {
 	testCase := testcases.NewABTestcase()
 	runIntegrationTest(t, testCase)
 }
+
+func TestIngestTypesTestcase(t *testing.T) {
+	testCase := testcases.NewIngestTypesTestcase()
+	runIntegrationTest(t, testCase)
+}
